#2374 change edit metadat to seperate component and cvreate a main page where you can toggle between edit and view +  fix some ui issues
